+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Other
 
+- Add PEP 620 header to Simple API HTML `PR #X`
 - Add S3 Docker Image building `PR #1092`
 - Move Docker containers to Python 3.10 `PR #1092`
 - Python 3.10 is now supported `PR #1073` -- Thanks **isidentical**